Corregidos modelos.

+++ Dviaje.Models/VM/ResenaCrearVM.cs
@@ -4,7 +4,7 @@
     {
         public string? Opinion { get; set; }
         public DateTime Fecha { get; set; }
-        public decimal Calificacion { get; set; }
+        public int Calificacion { get; set; }
         public int IdReserva { get; set; }
     }
 }
